Split graphql resolvers and typedef into their own folders
