+++ utils/apollo/schema/index.js
@@ -1,5 +1,5 @@
 import { makeExecutableSchema } from '@graphql-tools/schema';
-import { typeDefs } from './typeDefs';
+import { typeDefs } from './typedefs';
 import { resolvers } from './resolvers';
 
 export const schema = makeExecutableSchema({
