--- utils/apollo/schema/resolvers.js
@@ -1,134 +1,0 @@
-import User from 'models/User';
-import Space from 'models/Spaces';
-
-export const resolvers = {
-  Query: {
-    users: async (_, { userIds, email, name }) => {
-      // Fetch only the current session user by Name and Email
-      if (email && name) {
-        try {
-          const user = await User.findOne({ name, email });
-          console.debug('Fetching session user:', user);
-          return [user];
-        } catch (err) {
-          console.debug('Cannot fetch user', err);
-        }
-      }
-
-      // Fetches users that have an id found in userIds
-      const filter = {
-        _id: {
-          $in: userIds,
-        },
-      };
-
-      let users = [];
-      try {
-        users = await User.find(filter);
-        console.debug('Fetching users with specified userIds:', users);
-      } catch (err) {
-        console.debug(err);
-      }
-
-      return users;
-    },
-
-    spaces: async (_, { spaceIds }) => {
-      // Fetch all spaces
-      if (spaceIds?.length == 0) {
-        console.debug('Fetching all spaces:');
-        // populate('participants') fetches the corresponding User object for each object ID in the Space's participants attribute.
-        // Mongoose queries are not Promises by default. exec() turns the query into a Promise so we can use async/await after calling populate()
-        const spaces = await Space.find({}).populate('participants').exec();
-        return spaces;
-      }
-
-      // For Query by ID
-      const filter = {
-        spaceId: {
-          $in: spaceIds,
-        },
-      };
-
-      let spaces = {};
-      try {
-        // populate('participants') fetches the corresponding User object for each object ID in the Space's participants attribute.
-        // Mongoose queries are not Promises by default. exec() turns the query into a Promise so we can use async/await after calling populate()
-        spaces = await Space.find(filter).populate('participants').exec();
-        console.debug('Fetching spaces with specified spaceIds', spaces);
-      } catch (err) {
-        console.debug(err);
-      }
-      return spaces;
-    },
-  },
-  Mutation: {
-    createSpace: async (_, { input }) => {
-      const { name, description, userId, spaceId } = input;
-
-      let result = {};
-      try {
-        const space = new Space({
-          name,
-          description,
-          participants: [],
-          spaceId,
-          isActive: true,
-        });
-
-        result = await space.save();
-        console.debug('Created new space:', result);
-      } catch (err) {
-        console.debug('Cannot upload new space to database: ', err);
-      }
-      return result;
-    },
-    addUserToSpace: async (_, { input }) => {
-      const { spaceId, userId } = input;
-
-      let space = {};
-      try {
-        // Filters for the Space with the given spaceId
-        const filter = { spaceId };
-
-        // Updates by pushing userId into the participants array attribute
-        const update = {
-          $push: { participants: userId },
-        };
-
-        // Mongoose returns the space before the update occurs by default.
-        // {new:true} is the option argument that tells mongoose to return the space after the update occurs.
-        // populate('participants') fetches the corresponding User object for each object ID in the Space's participants attribute.
-        // Mongoose queries are not Promises by default. exec() turns the query into a Promise so we can use async/await after calling populate()
-        space = await Space.findOneAndUpdate(filter, update, { new: true }).populate('participants').exec();
-        console.debug('Updated space after populating participant field:', space);
-      } catch (err) {
-        console.debug('Cannot upload new space to database:', err);
-      }
-      return space;
-    },
-    removeUserFromSpace: async (_, { input }) => {
-      const { spaceId, userId } = input;
-
-      let space = {};
-      try {
-        // $pull removes the specified element from an array in mongoDB.
-        const filter = { spaceId };
-        const update = {
-          $pull: { participants: userId },
-        };
-
-        // Mongoose returns the space before the update occurs by default.
-        // {new:true} is the option argument that tells mongoose to return the space after the update occurs.
-        // populate('participants') fetches the corresponding User object for each object ID in the Space's participants attribute.
-        // Mongoose queries are not Promises by default. exec() turns the query into a Promise so we can use async/await after calling populate()
-        space = await Space.findOneAndUpdate(filter, update, { new: true }).populate('participants').exec();
-
-        console.debug('Updated space after removing a participant:', space);
-      } catch (err) {
-        console.debug('Cannot upload new space to database: ', err);
-      }
-      return space;
-    },
-  },
-};
