+++ utils/apollo/schema/resolvers/index.js
@@ -1,0 +1,6 @@
+import { Query } from './Query';
+import { Mutation } from './Mutation';
+export const resolvers = {
+  Query,
+  Mutation,
+};
